Work around bug getting rooms via V1 API.
When retrieving the list of available rooms in
`will.mixins.room.RoomMixin`, the api returns `Room` objects if the V2
api is used, but plain dicts using the V1 api.  To work around this, we
explicitly wrap the returned object in a new Room object, which works
fine for either type of object, as Room subclasses dict.